Move test_vm_scan to file with other VMs-related tests
